added column match with groups

+++ DataImporter/DataImporter.Worker/ReadFunctions/ExcelReadFunction.cs
@@ -1,6 +1,8 @@
 ﻿using DataImporter.Common.Utilities;
 using DataImporter.Importer.BusinessObjects;
 using DataImporter.Importer.Services;
+using DataImporter.Worker.FilePaths;
+using Microsoft.Extensions.Options;
 using OfficeOpenXml;
 using System;
 using System.Collections.Generic;
@@ -13,14 +15,16 @@ namespace DataImporter.Worker.ReadFunctions
         private readonly IDateTimeUtility _dateTimeUtility;
         private readonly IImportService _importService;
         private readonly IExcelDataService _excelDataService;
+        private WorkerFilePath _filePath;
         public ExcelReadFunction(IFileSearching fileSearching,
             IDateTimeUtility dateTimeUtility,IImportService importService,
-            IExcelDataService excelDataService)
+            IExcelDataService excelDataService,IOptions<WorkerFilePath> filePath)
         {
             _fileSearching = fileSearching;
             _dateTimeUtility = dateTimeUtility;
             _importService = importService;
             _excelDataService = excelDataService;
+            _filePath = filePath.Value;
         }
         public Guid UserId { get; set; }
         public int id { get; set; }
@@ -28,28 +32,40 @@ namespace DataImporter.Worker.ReadFunctions
         public List<List<string>> Excel { get; set; }
         public void ReadExcelData()
         {
-            string path = $"G:/Final/DataImporter/DataImporter.Web/wwwroot/Confirm";
+            //string path = $"G:/aspnetb5/DataImporter/DataImporter.Web/wwwroot/Confirm";
+            string path = _filePath.ConfirmPath;
             string s = null;
             var Files = _fileSearching.GetExcelFiles(path);
-
+            if (Files.Length == 0) Console.WriteLine("No file Exists");
             foreach (FileInfo file in Files)
             {
                 
                 s = file.FullName;
+                
                 string fileNamewithGuid = Path.GetFileNameWithoutExtension(s);
                 var fileNameSplit = fileNamewithGuid.Split('_');
                 var fileName = _fileSearching.GetFileName(fileNamewithGuid);
                 UserId = Guid.Parse(fileNameSplit[0]);
                 id = Convert.ToInt32(fileNameSplit[1]);
                 var groupName = _importService.GetGroupName(id);
-                var importid=_importService.GetImportId(fileName);
-                StoreExcelData(s,importid,id,UserId,fileName,groupName);
-                if (File.Exists(s))
+                var imports=_importService.GetImportList(fileName,UserId,id);
+                foreach (var item in imports)
                 {
-                    file.Delete();
+                    if (item.ExcelFileName == fileName && item.UserId == UserId
+                        && item.GroupId == id &&item.Status== "Pending")
+                    {
+                        StoreExcelData(s, item.Id, id, UserId, fileName, groupName);
+                        if (File.Exists(s))
+                        {
+                            file.Delete();
+                        }
+                        Console.WriteLine("Entry Done");
+                        _importService.Completed(item.Id);
+                    }
+                    else Console.WriteLine("Entry Already Completed");
+                 
                 }
-                Console.WriteLine("Entry Done");
-                _importService.Completed(importid);
+               
             }
         }
 
@@ -72,7 +88,10 @@ namespace DataImporter.Worker.ReadFunctions
                     string s = "";
                     for (int col = 1; col <= colCount; col++)
                     {
-                        s = s + worksheet.Cells[row, col].Value?.ToString().Trim() + '/';
+                        if (col < colCount)
+                            s = s + worksheet.Cells[row, col].Value?.ToString().Trim() + '/';
+                        else if(col==colCount) 
+                            s = s + worksheet.Cells[row, col].Value?.ToString().Trim();
                     }
                     list.Add(s);
                 }
